update statsclient
перенос resttemplateconfig в отдельный класс
